requirements.txt
streamlit>=1.35.0
plotly>=5.22.0
pandas>=2.1.0
requests>=2.31.0
anthropic>=0.40.0
python-dotenv>=1.0.0
python-dateutil>=2.9.0
